qr generating ready
