+++ software/v1-enlight/Makefile
@@ -5,7 +5,7 @@ SRC+=dia_configuration/dia_configuration.cpp dia_configuration/dia_screen_config
 SRC+=dia_functions.cpp dia_security.cpp dia_cardreader.cpp
 SRC+=dia_configuration/dia_screen_item_digits.cpp ./dia_screen/dia_int_pair.cpp ./dia_screen/dia_number.cpp ./dia_screen/dia_boolean.cpp
 SRC+=./dia_screen/dia_font.cpp dia_configuration/dia_screen_item_image.cpp ./dia_screen/dia_string.cpp ./dia_runtime/dia_runtime.cpp
-SRC+=./QR/qrcodegen.cpp
+SRC+=./QR/qrcodegen.cpp ./QR/EasyBMP.cpp
 SRC+=dia_configuration/dia_screen_item_image_array.cpp
 SRC+=dia_configuration/storage/dia_storage_interface.cpp dia_ccnet.cpp
 SRC+=dia_vendotek.cpp ./vendotek/vendotek.cpp
